+++ lib/domain/board/recruit/use_cases/write/validate_use_case_provider.dart
@@ -1,4 +1,4 @@
-import 'package:dongsoop/domain/board/recruit/usecases/validate_write_use_case.dart';
+import 'package:dongsoop/domain/board/recruit/use_cases/write/validate_write_use_case.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 final validateWriteUseCaseProvider = Provider<ValidateWriteUseCase>((ref) {
