+++ lib/domain/board/recruit/use_cases/write/validate_write_use_case.dart
